Feat: refactor banner management by removing ImageEditorModal, implementing BannerModal with enhanced link area management and validation
